BREAKING: In response wrapper, return data as **String**.
Rather than directly inlining the data's JSON.

This allows future implementations to seamlessly switch to other data
serialization formats, and adds some separation of concerns for
downstream parsing logic (i.e. easier to answer the question "was there
an error in parsing the server's response, or an error in parsing the
inner data?").